Criado feature de deletar e lista de vagas

+++ src/actions/delete_location_from_db.ts
@@ -1,9 +1,17 @@
+'use server'
+
 import { prisma } from "@/lib/prisma";
 
-export default async function deleteLocation() {
-    const location = await prisma.location.deleteMany({
+export default async function deleteLocation(id: string) {
+
+    if (!id) return
+    
+    const deleteVacancie = await prisma.location.deleteMany({
+
         where: {
-            id: 
+            id: id
         }
     })
+
+    return deleteVacancie
 }
